Добавляет оптимизацию
