+++ project/src/components/main-page-empty/main-page-empty.tsx
@@ -1,0 +1,22 @@
+import {useAppSelector} from '../../hooks/useAppSelector';
+import {getCity} from '../../store/action-process/selectors';
+
+function MainPageEmpty(): JSX.Element {
+  const currentCity = useAppSelector(getCity);
+
+  return (
+    <div className="cities">
+      <div className="cities__places-container cities__places-container--empty container">
+        <section className="cities__no-places">
+          <div className="cities__status-wrapper tabs__content">
+            <b className="cities__status">No places to stay available</b>
+            <p className="cities__status-description">We could not find any property available at the moment in {currentCity}</p>
+          </div>
+        </section>
+        <div className="cities__right-section"></div>
+      </div>
+    </div>
+  );
+}
+
+export default MainPageEmpty;
